refactor: rename styles files contact page
