+++ app/Home/components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import alan from '@page/assets/alan-buscaglia.png';
+import alan from '@app/assets/alan-buscaglia.png';
 import Image from 'next/image';
 
 import Controller from './Controller';
