+++ app/contact/components/Faq/components/ListItem/index.ts
@@ -1,1 +1,1 @@
-export * from './ListItem';
+export { default as ListItem } from './ListItem';
