--- app/contact/components/Faq/styles/styles.module.css
@@ -1,110 +1,0 @@
-.collapsible {
-  background-color: #777;
-  color: white;
-  cursor: pointer;
-  padding: 18px;
-  width: 100%;
-  border: none;
-  text-align: left;
-  outline: none;
-  font-size: 15px;
-}
-
-.active,
-.collapsible:hover {
-  background-color: #555;
-}
-
-.collapsible:after {
-  content: '\002B';
-  color: white;
-  font-weight: bold;
-  float: right;
-  margin-left: 5px;
-}
-
-.active:after {
-  content: '\2212';
-}
-
-.content {
-  padding: 0 18px;
-  max-height: auto;
-  overflow: hidden;
-  transition: max-height 0.2s ease-out;
-  background-color: #f1f1f1;
-}
-
-.accordion > input[type='checkbox'] {
-  position: absolute;
-  left: -100vw;
-}
-
-.accordion .content {
-  overflow-y: hidden;
-  height: 0;
-  transition: height 0.3s ease;
-}
-
-.accordion > input[type='checkbox']:checked ~ .content {
-  height: auto;
-  overflow: visible;
-}
-
-.accordion label {
-  display: block;
-}
-
-.accordion {
-  margin-bottom: 1em;
-}
-
-.accordion > input[type='checkbox']:checked ~ .content {
-  padding: 15px;
-  border: 1px solid #e8e8e8;
-  border-top: 0;
-}
-
-.accordion .handle {
-  margin: 0;
-  font-size: 1.125em;
-  line-height: 1.2em;
-}
-
-.accordion label {
-  color: #333;
-  cursor: pointer;
-  font-weight: normal;
-  padding: 15px;
-  background: #e8e8e8;
-}
-
-.accordion label:hover,
-.accordion label:focus {
-  background: #d8d8d8;
-}
-
-.accordion .handle label:before {
-  content: '\002b';
-  display: inline-block;
-  margin-right: 10px;
-  font-size: 0.58em;
-  line-height: 1.556em;
-  vertical-align: middle;
-}
-
-.accordion > input[type='checkbox']:checked ~ .handle label:before {
-  content: '\002d';
-}
-
-.accordion {
-  max-width: 100%;
-}
-
-.accordion .paragraph:last-child {
-  margin-bottom: 0;
-}
-
-.paragraph {
-  margin: 0 0 1em;
-}
