+++ app/contact/styles/contact.module.css
@@ -1,0 +1,99 @@
+/** Colors Palette **/
+
+.container {
+  --COLOR-PRIMARY: #36393f;
+  --COLOR-SECONDARY: #2f3136;
+  --COLOR-TERTIARY: rgb(32, 34, 37);
+  --COLOR-QUATERNARY: #292b2f;
+  --COLOR-QUINARY: #393d42;
+  --COLOR-SENARY: #828386;
+  --COLOR-TRASNPARENT: rgb(64, 68, 75);
+
+  --COLOR-WHITE: #ffffff;
+  --COLOR-GRAY: #8a8c90;
+  --COLOR-RED: #ff0606;
+
+  --COLOR-CHAT-INPUT: rgb(64, 68, 75);
+  --COLOR-SYMBOL: #74777a;
+  --COLOR-LINK: #5d80d6;
+
+  --COLOR-LIVE: #43b581;
+  --COLOR-NOTIFICATION: #f84a4b;
+  --COLOR-NOTIFICATION-DC: #dc274350;
+  --COLOR-DISCORD: #6e86d6;
+  --COLOR-MENTION-DETAIL: #f9a839;
+  --COLOR-MENTION-MESSAGE: #413f3f;
+
+  --COLOR-TUTI-PRIMARY: #000000;
+  --COLOR-TUTI-SECONDARY: #ea1889;
+
+  --COLOR-TUTI-FONDO-NAVBAR: #000000;
+  --COLOR-TUTI-NAVBAR-ICONS: #ea1889;
+
+  display: flex;
+  flex-direction: column;
+  margin-top: 80px;
+}
+
+.grid-contact {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 1rem;
+  align-items: center;
+}
+
+@media (min-width: 768px) {
+  .grid-contact {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 1rem;
+    align-items: center;
+  }
+}
+
+/** Contact Form section */
+.grid-contact__form {
+  grid-row: 2;
+}
+
+@media (min-width: 768px) {
+  .grid-contact__form {
+    grid-column: 1;
+    grid-row: 1;
+  }
+}
+.grid-contact__social {
+  grid-row: 1;
+}
+
+@media (min-width: 768px) {
+  .grid-contact__social {
+    grid-column: 2;
+    grid-row: 1;
+  }
+}
+
+/*  Social Media section
+ */
+
+.section-social__links {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+}
+
+/* Q&A section
+ */
+
+.section-qa {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+  order: 3;
+  border: 1px solid green;
+  padding: 40px 20px;
+  width: 90%;
+}
